% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AikenToBB.R
\name{AikenToBB}
\alias{AikenToBB}
\title{Create a Blackboard question bank from an exam in Aiken format}
\usage{
AikenToBB(input, title = NULL, rand.answers = TRUE, rand.questions = FALSE)
}
\arguments{
\item{input}{Path to the Aiken file}

\item{title}{Title for the question bank. If not supplied, the name of the file is used.}

\item{rand.answers}{Randomize the ordering of the answers?}

\item{rand.questions}{Randomize the question order?}
}
\value{
A list holding the question items. Also creates a zip file with the question bank.
}
\description{
This functions creates a Blackboard question bank from an exam in Aiken
format, which is a simple text format.
}
\details{
This is an old function that will be deprecated. Use \code{readAiken()} instead.
}
\examples{
exam <- AikenToBB("final.txt", rand.questions=TRUE)

}
