% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAiken.R
\name{readAiken}
\alias{readAiken}
\title{Read an exam in Aiken format into an R list format}
\usage{
readAiken(input, bestOf = 4)
}
\arguments{
\item{input}{Path to the Aiken file OR a character vector (one element per line)}

\item{bestOf}{Default: 4. Expect this many answers per question, and warn if fewer
(or more) are found.}
}
\value{
A list holding the question items. Each question item is a list with elements
\code{Q} (the question), \code{R} (the vector of responses), \code{correct} (the correct
answer, in a capital letter), and possibly \code{feedbackCorrect} (feedback text if the
answer is correct) and \code{feedbackIncorrect} (feedback text if the answer is incorrect).
}
\description{
This reads an exam in Aiken format into an R list,
suitable for further processing or export.
}
\details{
The following is expected in the Aiken file format:
\itemize{
 \item Questions are seperated by at least one empty line
 \item Questions can be across multiple lines; they are merged into a single line
 \item Answers start with a single letter (uppercase or lowercase) and a dot or parenthesis.
   A. or a. or a) or (a) are all valid.
 \item The letter prepending each answer is IGNORED. The answers are treated in order of appearance.
 \item Correct answer is indicated by ANSWER: followed by the correct letter. Case-insensitive.
 \item Comments start with \\ and are ignored
 \item However, lines starting with \\feedbackCorrect or \\feedbackIncorrect are NOT ignored.
   They are merged together into the feedback for a correct or incorrect answer respectively.
}

This behaviour ensures the maximum degree of tolerance in reading Aiken files.
An error is only generated if a section of the Aiken file has no question, or no answers,
or no correct answer (and the corresponding line is indicated).
}
\examples{
exam <- readAiken("final.txt")

}
