% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNLM.R
\name{readNLM}
\alias{readNLM}
\title{Read questions in NotebookLM JSON format into an R list format}
\usage{
readNLM(input)
}
\arguments{
\item{input}{The exported questions from NotebookLM. A JSON string, URL or file.}
}
\value{
A list holding the question items. Each question item is a list with elements
\code{Q} (the question), \code{R} (the vector of responses), \code{correct} (the correct
answer, in a capital letter), and possibly \code{feedbackCorrect} (feedback text if the
answer is correct) and \code{feedbackIncorrect} (feedback text if the answer is incorrect).
}
\description{
This reads questions in Google NotebookLM JSON format into an R list,
suitable for further processing or export.
}
\details{
The "rationale" for the incorrect answers is combined into a single feedback.
 The "hint" for each quesion is included as an extra comment.
}
\examples{
exam <- readNLM("questions.json")

}
