% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleExam.R
\name{shuffleExam}
\alias{shuffleExam}
\title{Shuffle the questions and/or items in a MCQ exam}
\usage{
shuffleExam(x, q = TRUE, a = TRUE)
}
\arguments{
\item{x}{An exam in list format, as read by \code{readAiken()} or \code{readNLM()}}

\item{q}{Shuffle questions? Default is \code{TRUE}}

\item{a}{Shuffle answers? Default is \code{TRUE}}
}
\value{
A list holding the shuffled question items.
}
\description{
Shuffle the questions and/or items in a MCQ exam
(in list format)
}
