% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAiken.R
\name{writeAiken}
\alias{writeAiken}
\title{Write a MCQ exam (in list format) to Aiken format}
\usage{
writeAiken(x, file = NULL, numbering = "[\%s] ", ignoreComments = TRUE)
}
\arguments{
\item{x}{An exam in list format, as read by \code{readAiken()} or \code{readNLM()}.}

\item{file}{The file to write to. If NULL, nothing will be written to disk.}

\item{numbering}{How to enumerate questions when exporting? Expressed as a format
string for \code{sprintf()}, default \code{"[\%s] "}. Set to \code{""} to disable numbering.}

\item{ignoreComments}{If \code{TRUE} (the default) no comments will be exported,
including feedback for correct or incorrect answers.}
}
\value{
A character vector (of length 1) containing the exam in Aiken format.
}
\description{
This takes an exam in list format and exports it to Aiken format.
}
