% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeKey.R
\name{writeKey}
\alias{writeKey}
\title{Output the answer key for an exam}
\usage{
writeKey(x, file = NULL, nrow = 12)
}
\arguments{
\item{x}{An exam in list format, as read by \code{readAiken()}}

\item{file}{The file name to write (with .csv or .xls or .xlsx
extension) or \code{NULL} if no file is to be output.}

\item{nrow}{The number or rows output in a key. Default is 12.}
}
\value{
The answer key in a matrix format (with \code{nrow} rows).
  At EUC we use this to copy the exam key to Scantron paper forms,
  for bulk machine-grading of exam papers.
}
\description{
This takes an exam in list format and outputs the annswer key,
in CSV or XLSX format (only if package WriteXLS is installed)
}
